Validators and bug fix (#26)
* Reorganize project

* Introduce validator logic

* Update and fix KU detection
