--- src/resolver/resolver.ts
@@ -1,11 +1,0 @@
-import { URL } from 'url';
-
-import { Message } from './message';
-
-export interface Resolver {
-  /**
-   * Resolve the request from the website
-   * @param url Url to resolve
-   */
-  resolve(url: URL): Promise<Message[]>;
-}
