+++ src/service/resolver/amazon/amazon-details.ts
