+++ src/service/resolver/scribd/scribd-information.ts
